Fixes #8: innodb.innodb test fails
Modified innodb.innodb and innodb_zip.innodb_16k tests to accept 513 as a
valid innodb_buffer_pool_pages_total value. It is valid when the OS page
size is 64K, in which case we round the real number of allocated memory
to 64K and thus allocate slightly more memory than with the default 4K
size.