Fixes #63: DebugPrintDeathTest.PrintEval failure.
Commit 5c2b61d checks the evaluation of args of DBUG_PRINT macro using
"division by zero" exception.

C99 says in 6.5.5p5: "... if the value of the second operand is zero,
the behavior is undefined". Also, according to the ARMv8 Architecture
Reference Manual if the divisor is zero, the result of operation
is zero and no exception is generated. Consequently, the
DebugPrintDeathTest.PrintEval unit test fails.

This patch avoid using "division by zero" exception.